chatbot_2/chatui_v1/requirement.txt
streamlit==1.38.0
requests==2.32.3
python-dotenv==1.0.1
httpx==0.27.0
